+++ packages/ts-newskit/src/slices/content-bucket-1/index.tsx
@@ -10,16 +10,15 @@ import {
   SliceHeaderProps
 } from '../../components/slices/slice-header';
 import { ArticleProps } from '../../components/slices/article';
-import {
-  LeadStoryDivider,
-  LeadStoryCell,
-  CellNoMargin,
-  CellWithCustomPadding
-} from '../shared-styles';
+import { LeadStoryDivider, StackItem, BlockItem } from '../shared-styles';
 
 import { CommentStack } from './comment-stack';
 import { ArticleStack } from './article-stack';
-import { CustomGridLayout } from '../shared/grid-layout';
+import { CustomStackLayout } from '../shared';
+import {
+  FullWidthBlock,
+  FullWidthHidden
+} from '../../components/slices/shared-styles';
 
 export interface ContentBucket1Props {
   section: SliceHeaderProps;
@@ -38,15 +37,34 @@ export const ContentBucket1 = ({
 
   const modifiedLeadArticle = {
     ...leadArticle,
-    contentTop: true
+    imageTop: breakpointKey === 'xs' || breakpointKey === 'sm' ? true : false
   };
 
   return (
-    <CustomGridLayout>
-      <CellWithCustomPadding xs={12}>
-        <SliceHeader {...section} />
-      </CellWithCustomPadding>
-      <LeadStoryCell xs={12} lg={9} xl={8}>
+    <CustomStackLayout>
+      <FullWidthBlock>
+        <BlockItem
+          $width={{
+            xs: '100%',
+            md: '720px',
+            lg: '977px',
+            xl: '1274px'
+          }}
+        >
+          <SliceHeader {...section} />
+        </BlockItem>
+      </FullWidthBlock>
+      <StackItem
+        $width={{
+          xs: '100%',
+          md: '720px',
+          lg: '760px',
+          xl: '840px'
+        }}
+        marginInlineEnd={{
+          lg: 'space060'
+        }}
+      >
         <Block>
           <Visible lg xl>
             <LeadStoryDivider
@@ -55,30 +73,51 @@ export const ContentBucket1 = ({
               position="right"
             />
           </Visible>
-          <LeadArticle {...modifiedLeadArticle} />
+          <LeadArticle
+            {...modifiedLeadArticle}
+            contentWidth={breakpointKey === 'xl' ? '312px' : '283px'}
+          />
         </Block>
         <Block>
           <Hidden md>
             <CommentStack comments={comments} />
           </Hidden>
         </Block>
-      </LeadStoryCell>
-      <CellNoMargin xs={12} lg={3} xl={4}>
-        <Hidden lg xl>
-          <Divider
-            overrides={{
-              marginBlock: 'space040',
-              stylePreset: 'dashedDivider'
-            }}
-          />
-        </Hidden>
-        <ArticleStack articles={articles} breakpoint={breakpointKey} />
-      </CellNoMargin>
+      </StackItem>
+      <StackItem
+        $width={{
+          xs: '100%',
+          md: '720px',
+          lg: '185px',
+          xl: '402px'
+        }}
+      >
+        <FullWidthHidden lg xl>
+          <FullWidthBlock>
+            <Divider
+              overrides={{
+                marginBlock: 'space040',
+                stylePreset: 'dashedDivider'
+              }}
+            />
+          </FullWidthBlock>
+        </FullWidthHidden>
+        <BlockItem>
+          <ArticleStack articles={articles} breakpoint={breakpointKey} />
+        </BlockItem>
+      </StackItem>
       <Visible md>
-        <CellNoMargin xs={12}>
+        <BlockItem
+          $width={{
+            xs: '100%',
+            md: '720px',
+            lg: '977px',
+            xl: '1274px'
+          }}
+        >
           <CommentStack comments={comments} />
-        </CellNoMargin>
+        </BlockItem>
       </Visible>
-    </CustomGridLayout>
+    </CustomStackLayout>
   );
 };