+++ packages/ts-newskit/src/slices/lead-story-2/article-stacks.tsx
@@ -1,19 +1,29 @@
 import React from 'react';
-import { Block, BreakpointKeys, Divider, GridLayout, Visible } from 'newskit';
+import {
+  Block,
+  BreakpointKeys,
+  Divider,
+  GridLayout,
+  Stack,
+  Visible
+} from 'newskit';
 import {
   LeadArticle,
   LeadArticleProps
 } from '../../components/slices/lead-article';
-import { FullWidthDividerMob } from '../../components/slices/shared-styles';
+import { FullWidthBlock } from '../../components/slices/shared-styles';
+import { LeadStoryDivider, RelativeBlockItem } from '../shared-styles';
 
 export const ArticleStack = ({
   verticalArticles,
   breakpoint,
-  horizontalArticles
+  horizontalArticles,
+  horizontalArticleContentWidth
 }: {
   verticalArticles: LeadArticleProps[];
   breakpoint: BreakpointKeys;
   horizontalArticles: LeadArticleProps[];
+  horizontalArticleContentWidth?: string;
 }) => {
   const modifiedVerticalArticles = verticalArticles.map(item => ({
     ...item,
@@ -27,7 +37,11 @@ export const ArticleStack = ({
     typographyPreset: 'editorialHeadline020'
   }));
   const articleGridVertical = (
-    <GridLayout columns={{ md: '1fr 1px 1fr' }} columnGap={{ md: 'space040' }}>
+    <GridLayout
+      columns={{ md: '1fr 1px 1fr' }}
+      columnGap={{ md: 'space040' }}
+      overrides={{ marginBlockStart: 'space000' }}
+    >
       {modifiedVerticalArticles.map(
         (article: LeadArticleProps, articleIndex, articleArr) => {
           const articleBorder = breakpoint !== 'xs' &&
@@ -44,7 +58,7 @@ export const ArticleStack = ({
           return (
             <React.Fragment key={article.headline}>
               <Block>
-                <FullWidthDividerMob>
+                <FullWidthBlock>
                   <Visible xs sm>
                     <Divider
                       overrides={{
@@ -53,7 +67,7 @@ export const ArticleStack = ({
                       }}
                     />
                   </Visible>
-                </FullWidthDividerMob>
+                </FullWidthBlock>
                 <LeadArticle {...article} />
               </Block>
               {articleBorder}
@@ -65,7 +79,11 @@ export const ArticleStack = ({
   );
 
   const articleStackHorizontal = (
-    <GridLayout columns={{ md: '1fr 1px' }} columnGap={{ md: 'space040' }}>
+    <GridLayout
+      columns={{ md: `${horizontalArticleContentWidth || '1fr'}` }}
+      columnGap={{ md: 'space060' }}
+      style={{ marginBlock: 'space000' }}
+    >
       {modifiredHorizontalArticles.map(
         (article: LeadArticleProps, articleIndex: number) => {
           const articleBorder = articleIndex !== 0 && (
@@ -77,18 +95,17 @@ export const ArticleStack = ({
             />
           );
           return (
-            <React.Fragment key={article.headline}>
+            <RelativeBlockItem key={article.headline}>
               <Block>
-                <FullWidthDividerMob>{articleBorder}</FullWidthDividerMob>
+                <FullWidthBlock>{articleBorder}</FullWidthBlock>
                 <LeadArticle {...article} />
               </Block>
-              <Divider
-                overrides={{
-                  stylePreset: 'lightDivider'
-                }}
+              <LeadStoryDivider
+                overrides={{ stylePreset: 'lightDivider' }}
                 vertical
+                position="right"
               />
-            </React.Fragment>
+            </RelativeBlockItem>
           );
         }
       )}
@@ -96,12 +113,19 @@ export const ArticleStack = ({
   );
 
   return (
-    <GridLayout
-      columns={{ md: '3fr 5fr ', xs: '1fr' }}
-      columnGap={{ md: 'space040' }}
+    <Stack
+      stackDistribution="flex-start"
+      flow={
+        breakpoint !== 'xs' && breakpoint !== 'sm'
+          ? 'horizontal-top'
+          : 'vertical-left'
+      }
+      spaceInline={
+        breakpoint === 'xs' || breakpoint === 'sm' ? 'space000' : 'space060'
+      }
     >
       {articleStackHorizontal}
       {articleGridVertical}
-    </GridLayout>
+    </Stack>
   );
 };