+++ packages/ts-newskit/src/slices/shared-styles/index.ts
@@ -1,41 +1,22 @@
 import {
-  Cell,
   Divider,
   getMediaQueryFromTheme,
   getSpacingCssFromTheme,
   styled,
-  Grid
+  Stack,
+  Block
 } from 'newskit';
 
-export const CellNoMargin = styled(Cell)`
-  & {
-    margin-top: 0;
-    padding: 0 16px;
-  }
-`;
-
-const setBlockMargin = (space: string) => ({ marginBlockStart: `${space}` });
-export const LeadArticleCell = styled(CellNoMargin)`
-  & {
-    padding: 0 16px;
-  }
-  position: relative;
-  ${getMediaQueryFromTheme('md')} {
-    ${getSpacingCssFromTheme(setBlockMargin, 'space040')};
-  }
-`;
-
-export const LeadStoryCell = styled(CellNoMargin)`
-  & {
-    padding: 0 16px;
-  }
-  position: relative;
-`;
-
+const setDividerPositionLeft = (space: string) => ({ left: `-${space}` });
+const setDividerPositionRight = (space: string) => ({ right: `-${space}` });
 export const LeadStoryDivider = styled(Divider)<{ position: string }>`
   position: absolute;
   top: 0;
-  ${({ position }) => getSpacingCssFromTheme(position, 'space000')};
+  ${({ position }) =>
+    getSpacingCssFromTheme(
+      position === 'left' ? setDividerPositionLeft : setDividerPositionRight,
+      'space040'
+    )};
 `;
 
 export const AvatarDivider = styled(Divider)`
@@ -44,17 +25,39 @@ export const AvatarDivider = styled(Divider)`
   }
 `;
 
-export const CellWithCustomPadding = styled(Cell)`
-  & {
-    padding: 0 16px;
-  }
-`;
-
-export const StyledLeadStoryCell = styled(CellWithCustomPadding)`
-  position: relative;
-`;
-
-export const StyledGrid = styled(Grid)`
+export const StackItem = styled(Stack)<{
+  $width?: { xs?: string; sm?: string; md?: string; lg?: string; xl?: string };
+}>`
   margin: 0;
+  position: relative;
+  width: ${({ $width }) => ($width && $width.xs) || '100%'};
+
+  ${getMediaQueryFromTheme('sm')} {
+    ${({ $width }) => $width && $width.sm && `max-width: ${$width.sm};`}
+  ${getMediaQueryFromTheme('md')} {
+    ${({ $width }) => $width && $width.md && `max-width: ${$width.md};`}
+  ${getMediaQueryFromTheme('lg')} {
+    ${({ $width }) => $width && $width.lg && `max-width: ${$width.lg};`}
+  ${getMediaQueryFromTheme('xl')} {
+    ${({ $width }) => $width && $width.xl && `max-width: ${$width.xl};`}
+`;
+
+export const BlockItem = styled(Block)<{
+  $width?: { xs?: string; sm?: string; md?: string; lg?: string; xl?: string };
+}>`
   width: 100%;
+  margin-inline: auto;
+
+  ${getMediaQueryFromTheme('sm')} {
+    ${({ $width }) => $width && $width.sm && `max-width: ${$width.sm};`}
+  ${getMediaQueryFromTheme('md')} {
+    ${({ $width }) => $width && $width.md && `max-width: ${$width.md};`}
+  ${getMediaQueryFromTheme('lg')} {
+    ${({ $width }) => $width && $width.lg && `max-width: ${$width.lg};`}
+  ${getMediaQueryFromTheme('xl')} {
+    ${({ $width }) => $width && $width.xl && `max-width: ${$width.xl};`}
+`;
+
+export const RelativeBlockItem = styled(Block)`
+  position: relative;
 `;