+++ packages/article/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.13.112](https://github.com/newsuk/times-components/compare/@times-components/article@7.13.111...@times-components/article@7.13.112) (2023-06-13)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
 ## [7.13.111](https://github.com/newsuk/times-components/compare/@times-components/article@7.13.110...@times-components/article@7.13.111) (2023-06-12)
 
 **Note:** Version bump only for package @times-components/article